Inital 20% solution of an RSS Feed Reader
It can read a feed; which is hardcoded in the source code. It accepts no
parameters and has nothing but some very basic functionality. It's about
ready to have some tests added to it.

It can view Description of an entry. It doesn't understand that most
entries are optional. It needs TONS of refactoring.

Testing:
  - None yet.