Add some tests, add POC ssh server functionality
We have some basic tests added to our model controller behaviors,
validating some of the behaviors present. We'll need to do more but
it serves as a start.

In addition to this, we implement the ssh server portion of the POC,
allowing users to connect via SSH in order to view the available
feeds.

More functionality coming.

Tests:
 - Actually wrote some tests, yay!